meta: patches: add path for sudo plugin
The sudo plugin seems not to be compatiable with the
newest release of dotbot. It also seems that the plugin
is not really maintained anymore, as there are still open
pull requests wich are long time ago are created. To not
fork the plugin, add a minimal patch to fix recoginize the
changes file structure of dotbot.

Signed-off-by: Oleg Karfich <oleg.karfich@wago.com>